move arry stuff into submodule
